prevent delete data mapper whilst job in progress
